refactor: storage utils in utils folder
